add some bounds check optimizations

+++ src/EventStore.ClientAPI.Abstractions/StreamPosition.cs
@@ -4,9 +4,9 @@
   public static class StreamPosition
   {
     /// <summary>The first event in a stream</summary>
-    public const int Start = 0;
+    public const long Start = 0L;
 
     /// <summary>The last event in the stream.</summary>
-    public const int End = -1;
+    public const long End = -1L;
   }
 }
